AmbiguousPathException is now in a separate file; PageNavigationService spacing & indentation fixed;
